refactor(insight): move implenmentation of  algorithms from insight module to data module
