+++ packages/ava/src/data/statistics/bayesian.ts
