--- packages/ava/src/insight/algorithms/category-outlier/index.ts
@@ -1,5 +1,0 @@
-import { IQR } from './IQR';
-
-export const CategoryOutlier = {
-  IQR,
-};
