+++ packages/ava/src/insight/algorithms/categoryOutlier.ts
@@ -1,0 +1,5 @@
+import { IQR } from '../../data/statistics/IQR';
+
+export const categoryOutlier = {
+  IQR,
+};
