--- packages/ava/src/insight/algorithms/change-point/index.ts
@@ -1,11 +1,0 @@
-import { bayesian } from './bayesian';
-import { windowBasedMean } from './window';
-import { pettittTest } from './pettitt-test';
-import { buishandUTest } from './buishand-u';
-
-export const ChangePoint = {
-  Bayesian: bayesian,
-  Window: windowBasedMean,
-  PettittTest: pettittTest,
-  BuishandUTest: buishandUTest,
-};
