--- packages/ava/src/insight/algorithms/change-point/types.ts
@@ -1,4 +1,0 @@
-export type ChangePointItem = {
-  index: number;
-  significance: number;
-};
