--- packages/ava/src/insight/algorithms/trend-direction/index.ts
@@ -1,5 +1,0 @@
-import { mkTest } from './mannkendall-test';
-
-export const TrendDirection = {
-  mkTest,
-};
