+++ packages/ava/src/insight/algorithms/trendDirection.ts
@@ -1,0 +1,5 @@
+import { mkTest } from '../../data/statistics/mannkendall-test';
+
+export const trendDirection = {
+  mkTest,
+};
